moved nft and balances from user
